extract gs1 list from list view
